requirements.txt
fastapi
uvicorn
sqlalchemy
fastapi_users_db_sqlalchemy
aiomysql